LayoutTests/fast/text/control-characters/complex-1-expected.txt

A control character should be rendered as a visible glyph. "ਅ\u0001ਆ" should be longer than "ਅਆ"

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.clientWidth is > ref.clientWidth
PASS successfullyParsed is true

TEST COMPLETE
ਅਆ
ਅਆ
